Remove old testing infra (#1759)
This removes the old testing infrastructure that was used to run tests inside the VS Code Extension Host. When we moved to vscode-test-cli the old test harness no longer runs, but the files were still present in the repository. This commit removes the files and the references to them.

It also removes the `test:unit:web` and associated test, as it was never improved to actually test anything. It was only there to test that the test harness was working, but that can wait until we actually have something to test in the browser beyond syntax highlighting.

This also removes the f-tfc branch from the test workflow triggers, as it was only used when we had the TFC feature branch in preview.

This updates the test workflow to run the tests using xvfb-run on Linux runners. This is necessary to run the tests on Linux runners as they require a display to run.

This comes from https://code.visualstudio.com/api/working-with-extensions/continuous-integration#github-actions

Removes macos-11 as it is EOL