Move wdio to src/test
This moves the WDIO tests to src/test/e2e. This is to make it easier to view and maintain all of the extension tests in the same place.

This required moving the wdio.conf.ts and the typescript files to src/test/e2e and updating the paths in the files to reflect the new location of the tests.

This also modified the npm prepare script to also install the wdio dependencies inside the src/test/e2e directory.

There were several minor fixes to existing WDIO tests to account for the recent HCP Terraform rename.

Note: You have to accept network connections for chromedriver to work. If this fails, then the wdio based tests will not run.