Store the hostname in the secret store
This commit stores the hostname in the secret store when the user logs in. This ensures that the hostname is available when the user logs in again.

This is necessary because the hostname is used to construct the API URL and the Web URL for a given TFE or HCP Terraform instance.