Change order of hostname quickpick options
This changes the order of hostname options in the quick pick by moving the new hostname option to the bottom of the list and presenting existing host names first. This makes it easier for users to select an existing hostname if they have one, as that is most likely the more common workflow.