Add TFE identification of instances
This change adds the ability to identify TFE instances by making a call to the `/ping` endpoint. This will allow the user to see the name of the instance they are connecting to in the quickpick menu.

There is still some hardcoding of 'HCP Terraform' in the code that needs to be updated to be more generic, but that can be handled in a separate set of work.