Provide a default formatter for Terraform Stack files
This change adds a default formatter setting for Terraform and Terraform Stack files. This will ensure that Terraform files use the correct formatter out of the box.

The configurationDefaults setting provides default values for settings that are not explicitly set in the user's settings. This setting is scoped to the specified language and does not affect other languages.

If the user already has a formatter set for Terraform files, this setting will not override it. This does not affect the user's settings, and the user can still change the formatter to a different one if they prefer.

This does not change when (onSave, onType, etc.) or how (formatting options, etc.) the formatter is applied. This only changes the default formatter that is used when the user does not have a formatter set.