Ensure date ordering for changie entries
This commit consolidates all prerelease entries into their related stable release entry file. This ensures that the changelog entries are ordered by date, with the most recent entries at the top.

The individual prerelease files are no longer necessary to keep around, as the information is now consolidated into the stable release entry.

The side effect of this is that we will need to do this same operation manually as we release a stable version as this is outside of changie's current capabilities.