Upgrade eslint to v9 flat config file
This upgrades eslint to v9 and switches to a flat config file. This is a breaking change for the eslint configuration. The new configuration is in eslint.config.mjs. This also updates prettier to work with eslint.

This results in an eslint configuration that is more in line with eslint and typescript recommendations, but results in a slighlty more strict ruleset being applied. This introduces some initial pain in fixing "working code" but should result in a more consistent codebase.