Use dependabot groups to group development dependencies
This adds groups to the dependabot configuration file for development dependencies.

There are many reasons to group dependencies, but the reason focused on here is optimizing for PR review. Instead of a single PR per dependency, a single PR with all the updates for a group is created. This is especially useful for development dependencies, which are often updated frequently and in small increments. This can lead to a large number of PRs being created. By grouping these dependencies, the rebase-review-merge-repeat cycle is reduced.

This commit introduces only groups for development dependacies in order to keep things simple. Likely we won't group production, but we can further refine the groups as needed.