Skip extension install UI tests for VS Insiders 1.97.0
In https://github.com/hashicorp/vscode-terraform/pull/1938 we discovered the failures seem to be related to https://github.com/redhat-developer/vscode-extension-tester/issues/1715 and VS Code Insiders 1.97.0. These failures may resolve with a new release of vscode-extension-tester or a new version of VS Code or they may not. We are skipping the tests for now to allow the dependency updates to merge and unblock future work